클래스 수정 이상함....살려주세요...

+++ src/main/java/com/knocksea/see/product/dto/request/ReservationRequestDTO.java
@@ -27,7 +27,7 @@ public class ReservationRequestDTO {
     private Long userId;
     // 상품번호
     private Long productId;
-    private int eduId;
+    private Long eduId;
     private Long reservationTimeId;
 
     public Reservation toEntity(ReservationRequestDTO dto) {
